# crypto/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Crypto currency app.",
        "title": "Go Crypto",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/SiHecate"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
        "/crypto/cryptoBuy": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Buy a specified amount of cryptocurrency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cryptocurrencies"
                ],
                "summary": "Buy Cryptocurrency",
                "parameters": [
                    {
                        "description": "Name of the crypto currency to buy",
                        "name": "cryptoName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Amount of crypto currency to buy",
                        "name": "amountToBuy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.BuyCryptoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
...
        "/crypto/cryptoSell": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sell a specified amount of cryptocurrency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cryptocurrencies"
                ],
                "summary": "Sell Cryptocurrency",
                "parameters": [
                    {
                        "description": "Name of the crypto currency to sell",
                        "name": "cryptoName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Amount of crypto currency to sell",
                        "name": "amountToSell",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SellCryptoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
        "/transaction/balanceTransaction": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of balance transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "List Balance Transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
        "/transaction/cryptoTransaction": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of cryptocurrency transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "List Crypto Transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
        "/user/addBalance": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a specified amount to the user's cryptocurrency wallet balance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add Balance to Crypto Wallet",
                "parameters": [
                    {
                        "description": "Amount to add to the balance",
                        "name": "add_balance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AddBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
        "/user/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the account balance of the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Account Balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.WalletResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.Status401Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status404Response"
                        }
                    }
                }
            }
        },
        "/user/cryptoWallet": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of cryptocurrencies in a user's wallet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List Crypto Wallet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.WalletListResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        },
...
        "/websocket": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetch cryptocurrency data from an external API and add/update it in the database using WebSocket.",
                "tags": [
                    "Cryptocurrencies"
                ],
                "summary": "Fetch and Add Crypto Data using WebSocket",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gokripto_controllers_crypto.Status400Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AddBalanceResponse": {
            "type": "object",
            "properties": {
                "available_balance": {
                    "type": "number"
                },
                "issuer": {
                    "type": "string"
                },
                "total_balance": {
                    "type": "number"
                }
            }
        },
        "controllers.BuyCryptoResponse": {
            "type": "object",
            "properties": {
                "amount_to_buy": {
                    "type": "number"
                },
                "crypto_name": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                },
                "user_balance": {
                    "type": "number"
                },
                "user_balance_after_buy": {
                    "type": "number"
                }
            }
        },
...
        "controllers.SellCryptoResponse": {
            "type": "object",
            "properties": {
                "amount_to_sell": {
                    "type": "number"
                },
                "crypto_name": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "total_profit": {
                    "type": "number"
                },
                "user_balance": {
                    "type": "number"
                },
                "user_balance_after_sell": {
                    "type": "number"
                }
            }
        },
        "controllers.Status401Reponse": {
            "type": "object",
            "properties": {
                "StatusUnauthorized": {
                    "type": "string"
                }
            }
        },
...
        "controllers.TransactionResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "type_info": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        },
...
        "gokripto_controllers_crypto.Status400Response": {
            "type": "object",
            "properties": {
                "StatusBadRequest": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}